requirements-test.txt
Pillow==7.1.2
ImageHash==4.1.0
matplotlib==3.2.1
